Add nmode_3_2 cuda kernel and tensorflow wrapper

+++ Kernels/add_one_op_kernel.cu.cc
@@ -2,14 +2,18 @@
 #define EIGEN_USE_GPU
 /* #include "third_party/eigen3/unsupported/Eigen/CXX11/Tensor" */
 
-__global__ void AddOneKernel(const int* in, const int N, int* out) {
-  for (int i = blockIdx.x * blockDim.x + threadIdx.x; i < N;
-       i += blockDim.x * gridDim.x) {
+__global__
+void
+AddOneKernel(const float* in, const int N, float* out) {
+  for (int i = blockIdx.x * blockDim.x + threadIdx.x;
+      i < N;
+      i += blockDim.x * gridDim.x)
+  {
     out[i] = in[i] + 1;
   }
 }
 
-void AddOneKernelLauncher(const int* in, const int N, int* out) {
+void AddOneKernelLauncher(const float* in, const int N, float* out) {
   AddOneKernel<<<32, 256>>>(in, N, out);
 }
